# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a Social Network - OnlyFans api specifications API REST .",
    "version": "1.0.2",
    "title": "Swagger Api-profile",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "yannick.simon@kysoe.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/country": {
      "post": {
        "tags": [
          "country"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "country post",
        "description": "country post desccription",
        "operationId": "countryPost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Country"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Country"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "country"
        ],
        "summary": "country list",
        "description": "country list desccription",
        "operationId": "countryList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Country"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/country/{country_id}": {
      "get": {
        "tags": [
          "country"
        ],
        "summary": "country get",
        "description": "country delete desccription",
        "operationId": "countryGet",
        "parameters": [
          {
            "name": "country_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "country"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "country delete",
        "description": "country delete desccription",
        "operationId": "countryDelete",
        "parameters": [
          {
            "name": "country_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/region": {
      "post": {
        "tags": [
          "region"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "region post",
        "description": "region post desccription",
        "operationId": "regionPost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Region"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "region"
        ],
        "summary": "region list",
        "description": "region desccription",
        "operationId": "regionList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Region"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Region"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/region/{region_id}": {
      "get": {
        "tags": [
          "region"
        ],
        "description": "region get description",
        "operationId": "regionGet",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "region"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "region delete description",
        "operationId": "regionDelete",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/city": {
      "post": {
        "tags": [
          "city"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "city post",
        "description": "city post desccription",
        "operationId": "cityPost",
        "requestBody": {
          "$ref": "#/components/requestBodies/City"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/City"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/City"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "city"
        ],
        "summary": "city list",
        "description": "city list desccription",
        "operationId": "cityList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/City"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/City"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/city/{city_id}": {
      "get": {
        "tags": [
          "city"
        ],
        "description": "city get description",
        "operationId": "cityGet",
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "city"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "city delete description",
        "operationId": "cityDelete",
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/tour": {
      "post": {
        "tags": [
          "tour"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "tour post",
        "description": "tour post desccription",
        "operationId": "tourPost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Tour"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Tour"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tour"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tour"
        ],
        "summary": "tour list",
        "description": "tour list desccription",
        "operationId": "tourList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tour"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tour/{tour_id}": {
      "get": {
        "tags": [
          "tour"
        ],
        "summary": "tour get",
        "description": "tour get desccription",
        "operationId": "tourGet",
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Tour"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tour"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tour"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "tour delete",
        "description": "tour delete desccription",
        "operationId": "tourDelete",
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "requestBodies": {
...
      "Country": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Country"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Country"
            }
          }
        },
        "description": "Country object that needs to be add",
        "required": true
      },
      "Region": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Region"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Region"
            }
          }
        },
        "description": "Country object that needs to be add",
        "required": true
      },
      "City": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/City"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/City"
            }
          }
        },
        "description": "City object that needs to be add",
        "required": true
      },
      "Tour": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Tour"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Tour"
            }
          }
        },
        "description": "City object that needs to be add",
        "required": true
      },
...
    },
    "schemas": {
...
      "Country": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Region": {
        "type": "object",
        "required": [
          "name",
          "country_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "country_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Region"
        }
      },
      "City": {
        "type": "object",
        "required": [
          "name",
          "region_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "region_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "City"
        }
      },
      "Tour": {
        "type": "object",
        "required": [
          "city_id",
          "profile_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "city_id": {
            "type": "string"
          },
          "profile_id": {
            "type": "string"
          },
          "time_start": {
            "type": "string"
          },
          "time_end": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Tour"
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}